requirements.txt
python-telegram-bot
SpeechRecognition
pydub
whisper
transformers
